requirements.txt
absl-py==0.12.0
anyascii==0.2.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
contractions==0.0.52
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
docopt==0.6.2
emoji==1.2.0
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
prompt-toolkit==3.0.19
protobuf==3.17.1
ptyprocess==0.7.0
pyahocorasick==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seqeval==1.2.2
silence-tensorflow==1.1.1
six==1.15.0
sklearn==0.0
sklearn-crfsuite==0.3.6
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
textsearch==0.0.21
tf2crf==0.1.32
threadpoolctl==2.1.0
tkintertable==1.3.3
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typeguard==2.12.0
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yarg==0.1.9
